fixed an alignment bug on sparc9

+++ src/dsm_array.c
@@ -28,7 +28,13 @@ typedef struct DsmConfig
 
 static DsmConfig *dsm_cfg = NULL;
 
-typedef int BlockHeader;
+/*
+ * Block header
+ *
+ * Its size must be 4 bytes for 32bit and 8 bytes for 64bit. Otherwise it could
+ * screw up an alignment (for example on Sparc9)
+ */
+typedef uintptr_t BlockHeader;
 typedef BlockHeader* BlockHeaderPtr;
 
 #define FREE_BIT 0x80000000
